rename :  Service layer 이름 변경 및 통일 (+패키지 변경)
